Stop snap iterating through empty vector
The snapping code has a hack-protection which attempts to stop snapping
between the boolean and it's original object. This code assumed that the
*it vector was always defined. Which for object snapping it is.

But for guide snapping it's not.

Thus lack of checking for the iterators existance caused a crash. This
check is correct because we know guides can't have boolean lpes.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3394