Update macOS build pipeline to v0.63
Apply updates from gtk-osx. This includes major updates like Cairo
1.16.0 to 1.17.6 and also a performance patch for Cairo that will
land in upstream shortly (upstream: MR 290).

Introduce building for arm64. Most of the preparatory work has
already been done before, so the final steps "only" require
updates to fontconfig, OpenMP and a special treatment for gsl
as we're trying to be more and more in-line again with upstream
gtk-osx.

Remove special treatment for building Python wheels.
This greatly simplifies the mibap build process by removing
a dedicated build step that was priviously run on GitLab (on an
older macOS version) in addition to the main build being run on
GitHub.

Add libwebp to enable exporting to webp. This is a dependency for
Pillow.

Refactor a lot of code. jhb takes care of the bootstrapping
process and now takes over the functionality that was previously
in toolset.sh.