Track tag name changes for Generic Ellipse
GenericEllise has an internal type, this is set at two points
during the sp-factor construction and again when writing out
the xml.

But undo and redo do not use the factory and obviously shouldn't
call the xml-write since it's an xml changing method already.
So to keep track it must track the tag name change event and keep
the genneric ellipse internal state consistannt.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/2879