Fix guide direction when snapping is on
There is a feature that changes the direction of a new guide
dragged from the ruler, so it's perpendicular to a path.

But because it was feeding the original normal into itself
the original orientation of the guide gets lost after the
first snapping has changed the normal.

We fix this by feeding in a copy of the original normal instead.

(cherry picked from commit d86167295f7e89731d2d0ab70730deee7c7885e5)