Re-enable the export button if canceled
If the prefs are canceled, then the export is not done
and it moves onto the next format in the batch export.

But because it continues, it skips cleanup steps and this
results in the export button being disabled.

This moves the cleanup until after all exports are completed.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/3605
(cherry picked from commit 6b9550fec84f0a103e7c1911cd89c28e6650fd6b)