Fix page trim and object align with the y-axis up
This is a partial backport of 6c652008085068284e and 1d6e8200896873d3f
with some additions suggested by Nathan Lee. When the y-axis points up,
ensure that objects are aligned to the visual top/bottom, depending on
user's action. When the page is fitted to the drawing, make sure pages,
guides, grids, viewport, and drawing content are shifted correctly, so
that the content lands on the resulting page, guides and grids are not
moved relative to it, and the view does not skip around.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3609
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3613