Update macOS build pipeline to v0.66
This brings back Ghostscript to the application bundle, providing
out-of-the-box support for EPS/PS files. The decision to bring it back
made under the premise that it can be easily removed again if we want
to pursue GPLv2+ compliance again.

Availability of Ghostscript automatically triggers more test in CI,
but four non critical tests are broken and have been disabled for
now (see 310-inkscape_test.sh).