Update macOS build pipeline to v0.69
This is a backport of all the advancements that have been made since
the release of Inkscape 1.2.1. While there's a lot of refactoring here,
nothing too fundamental has changed in how Inkscape or its depencies
are being built (incl. library updates), so we deem it safe to just
jump ahead to the most recent build pipeline instead of branching off
0.66 and backporting only a few select changes.

The most notable improvements are:
- CI for arm64
- make spellchicking possible
  (full fixing it requires a small additional change to
  inkscape-main.cpp, handled in a separate commit)