Fix image position on pixbuf input with y-axis up
When the y-axis points up, correctly fitting the document bounds to the
content requires knowing the current value of the document height. This
change ensures that the height is explicitly set when a new document is
created from a pixbuf, replacing the default value of 150px mandated by
the SVG standard.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3939