Implement context-aware actions in Layers dialog
Deletion and moving objects in the Layers and Objects dialog now work as
they used to in 1.1's Object dialog. This fixes a few UX issues, such
as those listed in https://gitlab.com/inkscape/inbox/-/issues/7792.

In the absence of any selected object, the Layers dialog now acts on the
active Layer (which is highlighted). Moreover, in Layers-only mode,
deleting using the Layers dialog deletes the layer (and not the selected
object).