dialog/objects: Centre blendModePopup() + cleanups
When reviewing the Controller MR, Mike was bemused that the blend mode
popover opens wherever the pointer was. It was already thus! But it is
fairly easy to always point to the middle of the icon, so do so; hence
it is consistent where we pop up, & we don't obscure the context icon.

* Factor out from getting the expander x to get_cell_area() -> rectangle
* Use it to get cell area for blend mode icon. Change to popup at middle
* Use Controller::has_flag() to more clearly check for pressed modifiers