Add stroke-width customizability for handles
Added the parser for `stroke-width` and changed the renderer
to use the property in the shapes construction.
Additionally, added the new implementation support to `SPKnot`
extending the project to now cover more tools and their different
states of handles.
Changed the initialisation method for `HandleStyle` to be more concise.

As part of the project, "Customizable Appearance of Canvas Controls"