Allow the resize handles to appear on white pages
When the desk color was dark and the page color is white, the handles must adjust their
pixels so they appear correctly. We do this by finding the location of the pixel and
where on the canvas it's going to fall and adjust which color it uses.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4513