Write start and end taper preferences on creation
The LPE params are not guarenteed to be consistant between their internal
memory data and their xml. So the attributes must be manually written out
after they have been changed.

This was not happening in this case.

Extra checks are also added to make sure there is no craash for broken files.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4462