Remove tool change and cope with double free in knot holder
We need to cope with the shape editor being destroyed first, or
the lpe being removed first. Either should be possible.

We also remove the tool changing code which is frightening. It
seemingly has no real effect on the functionality but if there is
regressions they must be fixed some other way.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4507