GradWStop: no Widget vfuncs; focus outline/tabbing
Be a GtkBox because that supports size-requesting via CSS, and do it
there, which is simpler & will not need adjusted when moving to GTK4.

Then means we can resolve the TODO re: no focus outline around the whole
widget when any stop has key focus, so do so. Then on trying this I felt
it underwhelming that Tab didn't move between stops, so this makes it so