Pass document into SPStyle for URI handling on merge
Any URI must have a reference to the document that the URI references
when doing a css merge in Object To Path, we forget to pass in this
document and cause it to lose the gradient uri reference.

The style code also needed to be ameneded since it didn't use the SPStyle
document, which is actually really important if the style doesn't have
an object yet because it's XML is promissed to be added in the future.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4460