Move the lpe warning to the LPE Dialog code
Before the ShowHandles code was removing itself if the user clicked cancel
this is a very dangerous way of asking the user to confirm the use of an
LPE, since the LPE has already been added and this is a singular and unique
confirmation only used for the show handles.

We remove this warning entirely on the collective decision that this is at
odds with how LPEs work.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4554