ObjectsPanel: Fix transparent hover on Alt, broken
by GTK3 EventControllerKey::modifiers not working properly. It seems
that only ::key-pressed|released get the correct state there, whereas
::modifiers gets some garbage that was inverting the intended logic.
Add warning comments in ui/controller.h to match; hopefully in GTK4 all
of that will become moot as *presumably* the controller Just Works there

Once that was fixed, it was pointed out that this was not updating the
highlighted item on hover anymore, so I fix that in the ::motion handler
– and THEN noticed it was not updating the highlighted state when the
Alt key was pressed/released, if the key-focus was not in our TreeView,
so this fixes that too, by listening for that key on the root window.

The logic from CommandPalette for listening for events on the root
window is hereby moved to ui/controller.h so that it can be reused, as
the key function is in the ObjectsPanel.