Finish integration of customizable canvas controls
- Outlines to handles are now completely implemented for all shapes
and settings. Its properties can be modified in the `node-handles.css`.
- Implemented the styling specified in ux#181 in css.
- Renamed type enum for controls and removed unimplemented
shapes, used these type values appropriately for different tools.
- Integrated all the different tools like measure, connector, gradient
and mesh. Guideline handles remain non-customizable through css
as they can be customized directly by the user anyway.
- Cleaned up any shape, fill or stroke setting code spread across the
codebase and replaced them with appropriate settings from the new
system.
- Changed `Handle` struct implementation, it now uses booleans
instead of any bit-flipping, made any needed changes as a result of it.
- Made `Property` struct more robust.
- Removed pixbuf/bitmap controls as they were unused, and as a
result unified the height and width in `CanvasItemCtrl` to be
reduced to a single variable.
- Still have an option in place to render ctrls according to code
if hard-coded style is a must for any handle.
- More code quality and efficiency improvements

As part of the project, "Customizable Appearance of Canvas Controls"