Render pen-tool anchor handles above anchor line
Fix the order between the pen-tool anchor
and the line displayed when drawing a curve.

Earlier the lines were displayed above the
handles, fixing that now they are rendered
under (before) the anchor handles.

Other merging related issues were alse resolved
in this commit. (merging with inkscape/master)

As part of the project, "Customizable Appearance of Canvas Controls"