dt-widget: Don't delete if set_desktop() with same
A thinko: in converting from raw new/delete to unique_ptr I omitted to
account for the case where the passed-in desktop pointer is the one we
already own, in which case we should not delete it, of course!

But as PBS pointed out, we can do better than comparing pointers: the
call from Desktop to DesktopWidget::set_desktop() seems to be pointless,
since the DesktopWidget creates the Desktop, so it achieves nothing to
say 'hey, please set your Desktop to the one you already created/have'.
So now we just remove all that cruft entirely.

Also, since the class is final, update `override` annotations to `final`

For me, this fixes:
https://gitlab.com/inkscape/inkscape/-/issues/4573#note_1547109655