desktop: Fix crash changing tool: reset/null order
As PBS pointed out, the change from raw pointer delete/=nullptr to
reset() would null the ptr before calling delete. But a Tool dtor may
call getTool(), so do it this way so that until dtor is done, our
getTool() will still return the valid ptr. This hopefully fixes:

https://gitlab.com/inkscape/inkscape/-/issues/4574#note_1549691674

A future direction could be to avoid dependency on this order and figure
out how to avoid the destructor triggering any calls to getTool() at all