DebugDialog: Remove as it's not showable!
https://gitlab.com/inkscape/inkscape/-/merge_requests/5789
 
As discussed with Tav on chat. It doesn't make much sense to, only on
Win32, redirect messages to dialog that no one ever attempts to show.
It seems the other, Messages dialog is the place these should go now,
and only if it's visible and ticked to log. So delete all of DebugDialog

>>>
dboles:
i guess it might change visibility of messages in the console on
Windows, but if so, we would want a proper fix for that, not 'dump
messages to a dialog no one can ever reveal'
so that removes the dialog, redirection to it, and the preference for same

Tav
Sounds good to me.
>>>