DialogMultip: Be Overlay+DrawingArea, not EventBox
For the MyHandle used to size the panes:
GTK4 won't let us use ::on-draw. So move to DrawingArea::draw, which in
GTK4 will be trivially replaced by DrawingArea.set_draw_func() – and
since even in GTK3, DrawingArea gets the events we need, drop EventBox
as that too will be gone in GTK4, so any less use of it is a benefit.

Also, make arguments const& where they can be, and unnamed where unused.

https://gitlab.com/inkscape/inkscape/-/merge_requests/5804