Fix Path-To-Stroke first node placement bug
In the specific case where the join type is join_pointy, the final
path segment is a line and the path is being closed during the
path-to-stroke operation; we move the start position to the very corner.

This removes the extra point that is always created when doing a stroke
to path on a rectangle, instead of 8 points, we had 9. Now the final
point and the first point are in the same position.