Incremental refactoring of extensions to move away from requiring the GUI
In this first step of refactoring we allow effect extensions to be called
with just a document, and without any GUI Desktop object.

We add the effect extensions to the doc action group, so they are available
directly without the app or window contexts for a future where we don't
need them.