InkScale: Fix bad clipped text @ widget !sensitive
When insensitive we would still draw the clipped label in white over the
normal text/fg color, which looks wrong if not sensitive: in dark themes
it means the bit of text up to the value is lighter than the rest, which
OK, not so bad – but in light themes like Adwaita it makes it invisible.

Fix this by just drawing the entire label in the normal theme_fg_color
when the widget is insensitive, just as GtkRange does not draw a fill in
that case. Also subtract clip_text_x from rect width to avoid overdraw
(unsure if this is a real risk, but subtracting seems obviously correct)

Also, apply an obvious optimisation: don't pointlessly create a layout
for empty string, and render nothing, if we don't have a label, which
seems to be by far the majority case!

Also, reuse alloc properties instead of getting from widget again, and
add missing include for Adjustment that doesn't error here but in GTK4.

https://gitlab.com/inkscape/inkscape/-/merge_requests/5817