AttrDlg|LPEItem|XMLTree|PageProps: menuize Popover
Now that the ContextMenu commit added menuize_popover(), we can use it
elsewhere, to make such popovers look more like GtkMenu did and hence
less different from how 1.3 looks and behaves. This also fixes the
tooltip on MenuButtons popping up over the popover after opening it...

While here, I make the XML Dialog layout menu close itself on selecting
an item, again because that was the old behaviour, and keeping it open
while it makes sense (since it immediately changes layout, so it's like
a 'live preview') is not useful enough I think to change the behaviour.

https://gitlab.com/inkscape/inkscape/-/merge_requests/5819