InkSpinScale: Use is_sensitive(), not StyleContext
We won't be able to use the StyleContext version in GTK4 `*`, nor need we.

Also, use the existing parent_type def when chaining up, not repeat it.

`*` Well, technically we might be able to since StyleContext is 'only'
entirely deprecated there, but since we know in advance that it'll be
gone in the future, we should take any easy change to avoid it.

https://gitlab.com/inkscape/inkscape/-/merge_requests/5824