Provide a test fixture for SVG-based tests
This commit provides a test fixture which makes it easy to create tests
based on SVG files containing test objects and corresponding reference
objects with specially formatted IDs.

To create such a fixture, inherit from the class provided here and set
up the test SVG file. The SVG file must contain pairs of elements which
represent a test object and a reference object for each test case.
For example, in the visual bounding boxes test, the test object's
bounding box will be compared to the rectangle given as a reference
object.

The advantage of arranging your tests in this way is that the SVG file
can be opened and inspected visually, making it evident that the tested
behavior is correct.