CanvItemCtrl: Fix inverted compare, C++20 warnings
set_pixbuf() was checking if the old and new values were *not* equal to
decide whether to exit early. It should only do so if they *are* equal.
This was confirmed with the previous author, PBS, in our Inkscape Chat.
Apparently this function is currently unused and will be removed later,
but until then or in case someone starts using it… letʼs make it right.

Alongside this I fix new C++20 warnings about implicit lambda captures.

See: https://gitlab.com/inkscape/inkscape/-/merge_requests/5856