IconRenderer: Remove unnecessary set_icon()→vfunc
The icon is set in the constructor and when the index property is
updated, so ensuring we set it on every render is simply wasteful.
We just need to ensure that add_icon() sets the name if needed (ie 1st).

After removing that, the vfunc just chains up, so just stop overriding.