GradientWithStops: Replace function not in GTK4...
...and in doing so, get tidier and possibly more efficient code by not
having to manually convert the pixbuf to a surface; let GDK handle that.

In my testing the drawing was the same before and after this change, but
making it will prevent it breaking build on GTK4, which lacks the old fn

Also, drive-by: #include and `using std::round` for unqualified round().

https://gitlab.com/inkscape/inkscape/-/merge_requests/5906