AttrDialog: Fix GTK4 build errors in self +depends
Most of the build errors in AttrDialog are results of better/stricter
conversions between TreeIter and TreeRow, and some wrong things we were
doing that ended up working because they were implicitly convertible in
gtkmm 3, but they're not anymore. So this cleans up all that. It also,
as stated, fixes build errors noticed along the way in depended-on files