Convert CanvasItemCtrl's _cache to image surface
Instead of storing the cache as an array of uint32_t, change it to a
Cairo::ImageSurface. This simplifies rendering it, and prepares for when
all handle types are rendered using Cairo. Also insert a null check to
avoid rendering from an empty cache.

Note that some complication arises from Cairo::RefPtr's refcounting
being thread unsafe; we have to use shared_ptr instead. This workaround
will disappear in Gtk4.