share/ui: Add .arrow to !recolor MenuButton arrows
Standard arrows like the `arrow` node in a `combobox` are not recoloured
but various manually created ones in GtkMenuButtons were, because they
lacked the `.arrow` CSS class I added to avoid that in some places.

Assuming we don't want these icons recoloured (as e.g. the +/- in spin
buttons also are not), this commit resolves that by adding said class.