LPEeditor: Optimise perf regressions in add_lpes()
We don't need a map to sort by label, instead let's just sort the passed
vector `lpes` in-place. To make this clear, move the sorting by category
AND label to this function. Then after, where needed, sort by label only

In addition, this also resolves that the loop on the `map lpesorted` was
copying the elements and then ALSO copying the pair element since it was
taking them by value, not by const& reference. We now use the latter.