Swatches: Reuse preview DrawingArea from statusbar - to preview palettes in popup menu
UX review and discussion:
https://gitlab.com/inkscape/inkscape/-/merge_requests/5976

The palette selector popup menu from the hamburger button at the bottom
of the window has a DrawingArea that previews the palette in each menu
item. This commit splits that to a separate ColorPalettePreview widget
and reuses that in the Swatches dialog too, so we have previews there,
which is something that Martin said:

> We really need some icons on those. Somehow.

The ColumnMenuBuilder gets a way to add a manually constructed MenuItem,
so we can create that in an exacting manner in the Swatches dialog. That
also sees refactoring to a to_palette_t(PaletteFileData) so that we can
reuse the logic of copying the name, id, and the now-needed colors when
using the _loaded_palette - as we now need the colors, to preview them.

This makes the ColumnMenuBuilder's max_width_chars option redundant now,
and meant that the ellipsize() function in the PopoverMenuItem needed to
be made available to the Swatches dialog, so those are removed and moved
to ui/util.h, respectively.

Finally, 1px spacing is added between the pre-existing menu's label and
the preview DrawingArea, and is replicated in the new usage by Swatches.