Fix regression to "always snapping" features
The always snapping features were removed during a previous refactoring
and need to be put back in order to not loose the functionality.

Because we are no longer using the tolerance to decide the always
snapping feature. We have to fix the historical bug in the line
snapper which depended on a large tolerance value.

When enabled, the grid and guide always flags force objects to ALWAYS be
snapped to those things.

The object, align and distribute always flags do not behave the same
as the grid and guides snappers so their preferences have been labelled
differently to make this clear. They act as exceptions to soften the
always flags.

Guides are protected from always snapping to themselves

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4510