Revert "display-popup.glade: Donʼt use GtkModelButton, as…"
This reverts commit d620ab8ce5115154473df1b17d145ed10d3aaa73.

See https://gitlab.com/inkscape/inkscape/-/issues/4646 and new comment.

Changing the action's target value externally (not by these buttons) -
while having them in a group - also causes toggled/click signal of the
radio button and for the previous value... which then activates that
action target value, which deactivates the 'new' one, etc etc. In GTK4
groups and actions can coexist, but in GTK3 we are not so lucky. You'd
think we could just remove the group, but then all buttons are always
active, not only the right one. So we'd need another bodge for that...

I have a personal project using integer radio actions and `:group` in
GTK4, and that works fine, so rather than trying to find an intervening
bodge that will work on GTK3 and GTK4, let's just revert the change
until we're on GTK4, at which point it should be re-apply-able and work.