Port Controller|Desktop/Events|InkscapeWin|Cursor
This was basically an attempt to get the window state stuff in inkscape-
window and desktop migrated, and as things do, it grew arms and legs as
I worked through the compiler errors. So this basically gets Controller
ready for GTK4 with minimal changes, migrates that stuff in Desktop and
the InkscapeWindow, updates cursor-utils for GTK4 and to add a function
that both loads and sets a cursor on a widget instead of being required
to type the verbose `widget.set_cursor(load_svg_cursor(widget, ...))` …
…updates Gtk::GestureMultiPress to GestureClick, Gdk::Window to Surface
and various enums related to the Controller module and Gdk::Device/Tool,
as well as removing Controller faffs that have proper provisions in GTK4
– and the manage() stuff because that too was just a workaround for GTK3

I make no claims this is in any way complete, but it's decent progress.

ui/cursor-utils: Update & workaround for GTK4

* Don't include all of <gtkmm.h> in the header, just what we need
* Replace Gdk::Window with Gdk::Surface
* Gtk::IconTheme now takes a Display, not a Screen
* Gdk::Cursor can't be created from Cairo::Surface anymore, and nor can
  Gdk::Pixbuf, so we seem limited to using Gdk::Pixbuf directly. Time
  will tell if that ends up looking worse as the comment said, then we
  can revisit.