InkscapeWindow: Fix tracking state of toplevel to…
…not require that a toplevel always exist for this widget, which it does
not during close. Just binding a ref to the original toplevel did not
work for me either as ASAN diagnosed a use-after-free. So just do this:
always try to get the toplevel anew to check its state, and if we do not
have one anymore, act like the state has changed from `_old` to empty {}