Shortcuts: Fix access @ destroyed Gtk::Application
Perhaps due to some change in GTK/mm4, keeping the app as a RefPtr (in
gtkmm4, now just a std::shared_ptr) meant our destructor tried to unref
it, which failed. We have no business doing that anyway as we just want
to observe the application, so just keep a raw, non-owning/sharing ptr.