Manage batch export filenames better
This improvement splits out the batch export path and the filename
components into two widgets and two fields. It also makes the data
stored in the document specific to the batch exporter and not shared
with the single exporter.

A new checkbox allows the user to decide if they want to overwrite
existing batch filenames. It's switched on by default.

The path is stored as a relative filename from the document's path
this allows a consistant reuse of the export location.

The way relative paths are calculated has been changed. By default
if there are more than two parents away, it will now return the
absolute path instead of a pile of ../../../ etc.

Fixes: https://gitlab.com/inkscape/inbox/-/issues/6983