Fix a leak in the Extension system
The extension objects, which used to be carelessly allocated on the
heap and left the to leak are now owned by the "Extension database".

The collection of owning handles to extensions added to the database
is separate from the map used to look them up, which is left unchanged,
except for a corner case in which the database could previously free
an extension.