Remove non-owning views
The extension database now holds owning unique_ptr instances for
all extensions. The "register" and "unregister" functions have
been modified; it's no longer possible for an extension object to
exist if the extension has been disabled.

As a follow-up to the refactor of the memory management in the
extension system, small code quality issue and documentation
discrepancies are addressed.

Co-authored-by: Thomas Holder <thomas@thomas-holder.de>