Restore toolbar collapse logic
Reimplement the toolbar collapsing using Bin.

The code closely follows the original, with the exception that we cache
the width decrease when collapsing and re-use it when deciding whether
to expand. This is to prevent oscillations caused by the calculated and
actual width increase not quite matching, due to a combination of
various spacing factors.

As a result, the window is no longer huge and bigger than the screen.