Fix sensitivity of Select Toolbar widgets on startup with new document
Some widgets in the Select Toolbar are disabled or enabled depending
upon the selection being empty or not. This is handled by connecting to
the Selection changed signal. The connection was being made in the
toolbar constructor but being unconnected in the on_unrealize()
function. Starting up Inkscape with a new document causes two cycles of
on_realize()/on_unrealize() which deleted the connection (this doesn't
happen when starting up Inkscape with an existing document), leaving the
toolbar unable to respond to changes in the selection. By moving
connecting to the Selection changes signal to the on_realize() function,
the signal is always properly connected.

I'm not sure why one wants to unconnect the signal in the on_unrealize()
function. If it is to avoid updating the widgets when the toolbar is not
visible, this doesn't work as on_unrealize() is not called when the
toolbar is hidden (i.e. when switching to a new tool). If this was the
intent, then using on_map()/on_unmap() would be the correct place to
connect/unconnect to the signal. Otherwise, disconnecting in the toolbar
destructor would be better. Fixing this is out of scope for this commit.