New optinal inx key: show-stderr (for custom guis)
If this key exists and is set to true stderr
is also shown for extensions having a
custom gui. If it does not exist for such
extensions it is assumed to be set to false.

The key is ignored if set for extensions without
a custom gui.

Relates to https://gitlab.com/inkscape/inbox/-/issues/9679