Fix some issues with Path > Flatten when SPImage is used
Firstly we make sure that the remove preference isn't being used
when we call the flatten or fracture actions. The remove preference
is specific to the shape builder tool.

Secondly we make sure that the path vector is complete. SPImage path
vectors were not closed objects, so the flatten would just delete the
objects instead. We actually want it to attempt to flatten as much
as possible, even if there are unclosed lines.

Thirdly we convert all strokes-to-path automatically to preserve
the stroke's visual appearance and use it in the flattening.