Do not leak Extension implementations
The implementations of extensions have a dual ownership model: either
they are created and owned by Inkscape, or they are owned by an
external library. To prevent memory leaks, a helper class is provided
which manages this dual ownership situation; an Extension object now
holds an instance of this class.

When building extensions from memory, we create owning pointer to these
extensions implementations by calling std::make_unique(). In fact, the
migration from operator `new` to `std::make_unique()` is responsible for
most of the diff in this commit.