Revert MR6003 and fix export path generation
The previous Merge Request did not account for the variation in input
and led to the blind concatenation of paths. Each selection or change
of settings would append yet another full path to the field.

This fix reverts 1858aeaf and fixes the issue with better control of
the filename. We recognise that any call to "getExportFilename" may
return a full path or a relative path depending on what is stored in
the xml. We detect this in the renamed function "prependDirectory"
which adds on the directory needed for this path.

Also fixed is the saving of multi-page filenames when more than one
page is selected the filename should be saved against the document
and not the selectedPage which is no longer used in the export dialog
to decide on the "selected page" for export.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/4695