Fix updating of dash pattern Gtk::Entry
The code tests if the dash pattern widget has focus to prevent the
cursor from being placed at the beginning of the entry. In Gtk3,
the Gtk::Entry would have focus, in Gtk4, it is the Gtk::Text inside
the Gtk::Entry that has focus (by design). A new function is added
that tests if any widget or child of that widget has focus.

The focus test should not be necessary, however, the StrokeStyle widget is
needlessly updated a second time via StrokeStyle::selectionModifiedCB().

This commit also ensures that the dash pattern entry is updated
properly everytime the selection changes.