Implement relative subdivision for seamless reunions
Add a new option to Path::ConvertWithBackData() to instruct that
subdivision should be performed relative to the size of each path
segment, rather than to an absolute length scale.

Make use of this in path-boolop.cpp to ensure seamless reunions in the
shape builder / boolean operations / boolean LPE. That is, the result of
fracture + union should not contain artifacts. Currently this is still
not guaranteed, but the situation is at least improved.