Initialize toolbars to nullptr in header file
When the toolbar is not initialized to nullptr in the header file, the   
initial call of set_icon_sizes to the toolbar is from the constructor    
before the toolbar pointers are initialized causing a crash.             
Initializing them to nullptr when they are declared in the header file   
prevents this and fixes the crash.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4757