Use a caching key for items at point cache
When constructing a cache, we need to be sure that the cache uses a key based on the
inputs to the function where caches are constructed. Before there was only a single
cache and this is incorrect given the variables it's posible to call the function with.

Now we have a caching key, the self-calling function was made internal and not a class
function, a new function called get_flat_item_list manages the cache internally instead
of depending on the caller to do cache management as well as starting at the root node.

The boolean flag for detecting a valid cache is removed in favour of checking the size
of the mapped item list. Clearing all and any caching keys is done with clear()

The use of a backup item list in getItemAtPoint is no longer needed as this was trying
to cope with the lack of a caching key. This is now greatly simplified.

Finally we add a new boolean variable to pick hidden and locked objects and let
the contet menu use this new argument. The returned values are still cached, but not
in the same cache as other calls.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4777