Fix clipping paths when used with groups in the shape builder
The shape builder allows the modification of clipping paths for
images and images within clones. But if the image selected has
the clipping path attached to a parent group (that is selected)
the clipping path will instead be ignored and destroyed.

Affects: Shape Builder (tool), Flatten, Fracture (actions)

A new clipping path vector getter is added to sp-item and a way to
combine them is added in path-util so paths can be combined by
other future functionality too (for example deep ungrouping)

Some additional adjustments to the booleans builder allows it to be
run without an SPDesktop in the case it's being run from a commandline.

A test is added for sp-item, which tests the new general functionality
but not the shape builder itself as there aren't good ways to test that.

Co-authored by: pbs