Revert event controller use in canvas
Revert the switch to event controllers in the canvas, since they're only
needed for porting to GTK4, whereas the next release is on GTK3 where
the event controllers only cause regressions.

Remove the redundant last_mouse tracking by the desktop, and guard
against the mouse not being over the canvas during a pinch zoom gesture.