Invalidate node cache when changing layer
Because the data in the node cache is dependant on the selected layer
when selecting outside of groups. The cache must be invalidated when
the layer changes as groups (which are not layers) that the use has
entered are considered layers during that time.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/6176