Improve UX by disabling color selection arrows at bounds
Changed the color selection arrows to become insensitive (disabled)
when the user reaches the start or end of the color palette.

This visual feedback informs the user that no more scrolling is
possible, which aligns with common UI patterns and enhances the user
experience by providing immediate and clear feedback on the
navigation limits.

Resolves: https://gitlab.com/inkscape/inkscape/-/issues/4114