[Corners LPE 3] Alert changes. This MR keep track on path changes and apply a uniffied function to the LPEs to store from LPEItem:
Number of shapes (child shapes for groups in the future read comment).
Number of curves
Number of subpaths

Storing this value we can call a update when any of this value is changed.
Because need computed time we only run over 2 specific LPE that current use. 
Is easy apply to more just set a variable in the LPE constructor.