Restoring Minwaita look and feel resembling v1.3
There is no gtk-4 specific Minwaita theme in our sources
and perhaps we don't have to maintain our own theme.
Minwaita-Inkscape.css contains minimal set of changes
that can be applied on top of default Adwaita theme.
This is an attempt to make minimal changes to restore look
and feel of Inkscape v1.3 when selecting Minwaita theme
from preferences.