Allow internal tool shortcuts to be customized
To do this we make use of the functionality in ActionAccel which Rafael had
used in the pen tool to fix the undo/redo accelerator code. Firstly we copy
that fix into the booleans-tool which also has a custom undo/redo action and
next we make tool specific accelerators by adding the metadata without the
actions allowing them to be customized in our existing UI but not cause any
of the side effects that creating actions would cause.

Because ActionAccel objects use the Shortcuts object on creation, we needed
to fix the Shortcuts initialization as it caused a race condition between the
tool initialization and the window actions initialization. But since ActionAccel
doesn't need any shortcut initialization to work, we turn init off for that
constructor.

Because tools can have their own contexts, we adjust the preferences to test
the context of actions before conflicting with them. Allowing two different
tools to use the same shortcut key.

Also fixed is the use of the action label, where possible, for the warning msg.

The default keys are stored in inkscape.xml

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4318