Startup screen template list: Avoid loading icons twice
Rendering the SVG icons of templates is a huge bottleneck in the startup time. (ca. 1-2 seconds).

Cut this startup time in half by rendering each file only once and then caching the result. This saves ca. 0.5-1 second of startup time because currently we have ca. 90 templates but only 37 distinct icons.