Share BatchItem management between single and batch exports
Because export-single uses the BatchItem and it's infrastructure
for previewing the multiple pages in a multi-page format such as pdf
we want to have the same management of BatchItems shared between
the two implementations.

This is probably not the final form that this code should get to
but this should move it in the right direction of a more dedicated
MultiItem preview utility in a future refactoring.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4523
Fixes https://gitlab.com/inkscape/inkscape/-/issues/4524