Single page must be added more carefully in export
There are two fixes here, firstly the viewport check is changed so
that it compares the floor of the corner of the page coordinates
as the crash was caused by a tiny offset in the viewbox.

Secondly, the export itself shouldn't use the viewport page anyway
since it's already tested if there's one page, it can just add it.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4747