PDF export: remove "unsupported target 0" warnings
Remove useless warnings about "unsupported target 0" printed on the ter-
minal during PDF or PostScript export. These warnings are harmless and a
result of a developer misunderstanding.

To be precise, Cairo allows adding metadata to PDF and PS surfaces only.
The PDF and PS renderer in Inkscape can however create image surfaces as
well, usually to apply masks and clips, at intermediate stages of render
operations. Since the internal functionality of creating a surface will
attempt to add metadata by default, a warning is printed saying that the
"image" surface type does not support it. The user needn't know about
this fact, since we are not trying to add metadata to an image surface
and the rendering code is in fact working as intended.