Fix gradient color drag undo issue in F&S dialog
This commit ensures that the changes made to the gradient stop color
through the Fill & Stroke dialog are pushed into the undo stack only
when the drag action is complete and the user has released the dragger.
This prevents pushing all the values changed during the drag action
into the undo stack, improving the user experience.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/4162