Fix grid alignment on paste and missing grid undo
We want to round to the pixel grid before asking for document's own grid alignment.

Also fixed was the missing undo for all the grid settings, a new undo_id setting
is added to the undo enabler so we can ask for maybeDone instead of done.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4827