Update grid widget removal for GTK4 compatibility
There was an issue with removing grids in Document Properties >
Grids > Remove. On clicking Remove, the widgets were not being
deleted and on subsequent interaction with the deleted grid
widgets, there was a crash. This patch fixes that issue.

The removal process for Grids has been adapted to align with
the Gtk::Notebook changes in Gtk4. Previously, the code was
not functioning as widget traversal logic was based on GTK3's
methods. The updated code now makes use of for_each_page method
from ui/util.h to remove the grid widgets