Fix min size of messages+selectors dialog
Changed the place of minimum size request of messages
and selectors dialog to their content-displaying children.

This is done to still ensure they are wide enough for the
content to be legible but now also allow resizing of the
sidebar below the minimum width of the contained dialogs,
should the dialog be closed.

This was not possible before and mirrors the behaviour of
other dialogs with a non-standard min. width.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2733