Fix crash when adding/deleting pages from toolbar
As a part of the GTK4 migration, the addition of window-level actions
was moved to an earlier stage of InkscapeWindow constructor, at which
point the desktop widget and _desktop member has not yet been set to
point to the SPDesktop. This resulted in the capture of null pointer
by window-level actions that nonetheless queried the window before it
was fully initialized.

The fix, for now, is to capture the window pointer (since these are
window-level actions) and query the window only once the action runs,
which can only happen after the window has been fully initialized.

This fixes segfaults when pressing the buttons to create or delete a
page in the toolbar of the Pages tool.